Add support for scram-sha-256 password authentication.
The module was uses hardcoded MD5 password encryption. [1]
(If encrypted is None, True, or md5, it will use PostgreSQL's MD5 algorithm.)

In Postgres 15 will show warning about the md5 usage in pg_hba.conf.
scram-sha-256 is the currently method to store passwords. [2]

To keep backwards compatibility I change the encryption field only, if the user
is configured to use scram-sha-256.

[1]: https://docs.saltproject.io/en/latest/ref/states/all/salt.states.postgres_user.html
[2]: https://www.postgresql.org/docs/current/auth-password.html